dockerized build
