+++ Dockerfile
@@ -1,0 +1,21 @@
+FROM python:3.10
+
+RUN pip install poetry==1.7.1
+
+ENV POETRY_NO_INTERACTION=1 \
+    POETRY_VIRTUALENVS_IN_PROJECT=1 \
+    POETRY_VIRTUALENVS_CREATE=1 \
+    POETRY_CACHE_DIR=/tmp/poetry_cache
+
+WORKDIR /bot
+
+COPY pyproject.toml poetry.lock ./
+RUN touch README.md
+
+RUN poetry install --no-root && rm -rf $POETRY_CACHE_DIR
+
+COPY src ./src
+
+RUN poetry install
+
+ENTRYPOINT ["poetry", "run", "python", "src/bot.py"]
