+++ pyproject.toml
@@ -1,9 +1,8 @@
 [tool.poetry]
-name = "spritebot"
+name = "bot"
 version = "0.1.0"
 description = ""
 authors = ["Joanie <greystorm101@gmail.com>"]
-readme = "README.md"
 
 [tool.poetry.dependencies]
 python = "^3.10"
